Major changes
As I don't totally understand the implementation, I probably introduced some inaccuracies and grammar issues.  But I'm trying to make it easier to understand, so please try dumbing down explanations that I misunderstood : )

One major change was switching SPV36 to SPV-C (see https://github.com/hlandau/ncdocs/issues/1) but I did not rename the other modes.

Instead of saving the naming for UTXO CB and NX CB for one big paragraph at the end, I tried to set up a pattern in which the basic concept is introduced and then given a name.  

I also removed the paragraph dedicated to explaining the choice of block depth for attestation and inlined the block depth number elsewhere.  It felt like that paragraph introduced a complex topic that was then immediately solved.  Remember that this is a blogpost, we can leave details regarding requirements for potential alternative implemenations in the NEP.